requirements.txt
attrs~=20.3.0
numpy~=1.20.1
pytest~=6.2.2
setuptools~=49.2.1